Minor fixes for semantic manifests in test cases (#1818)
This PR makes a couple of fixes to the semantic manifests used in test
cases.

* In `simple_manifest`, `booking_value` is used as a metric, but it
wasn't defined as one.
* There are two semantic manifests `multi_hop_join_manifest` and
`partitioned_multi_hop_join_manifest` that are supposed to be the same,
except the latter has a partition dimension in the models. It seems they
have diverged, so additional elements have been added.