Improve engine initialization speed via semantic graph (#1835)
This PR removes the option to use the legacy resolver when initializing
the `MetricFlowEngine` / `SemanticManifestLookup` in favor of the
semantic-graph-based resolver. For complex semantic manifests, this
improves initialization times by >100x for complex manifests.

The new resolver includes a fix for detecting ambiguous join paths. For
such detection, a semantic model with a `UNIQUE` entity should be
considered as a candidate model for joining on that entity. Previously,
it was not, and the generated query may use any of the possible join
paths and give inconsistent results. The fix detects the ambiguous path
in such queries and raises an error.

As the SG resolver returns a different implementation for result
objects, there are a number of snapshot updates.