Remove version requirement for dbt adapters in `dbt-metricflow` (#1837)
The `dbt-metricflow` package specifies extra dependencies e.g. `pip
install dbt-metricflow[dbt-duckdb]` that installs the compatible dbt
adapter for use with the CLI. Since the adapter only need to be
compatible with the version of `dbt-core` used by MF, we can remove the
version requirement and `pip` should install the latest version that is
compatible with the required version of `dbt-core`.