Pin to a non-dev version of DSI for release (#1854)
DSI was pinned to a dev version in the last release, which is a problem
for some users who use python package managers that ignore dev versions.
This pins it to the prior valid version so we can cut a new release
pointing to that.

Also pins click due to a bad release that went out today and caused
issues building hatch with python 3.12.