Bug fix: don't error if there are no time dims in manifest (#1862)
This was already approved, but I rebased it on top of my stack due to
some of the CI version resolutions that were needed to unblock CI.
Something weird happened where it automatically merged into the base
branch at that point, so I needed to create a new PR for this.
TLDR; this bug fix was already approved but needs re-approval.