Bump to new DSI dev version (again) (#1863)
We previously upgraded to this version of DSI, then reverted back to a
non-dev version for release. Now that the release is finished, we should
again use the newer version of DSI. This is needed before releasing a
new version of MF to the server. This is because mantle might be
producing manifests with the new fields included in the new DSI version,
and MFS needs to ensure it can consume those fields.