Use `node_relation` to render time spine table SQL (#1875)
A
[customer](https://dbt-labs.slack.com/archives/C05FWBP9X1U/p1759251652128099)
is running into a bug in Big Query because we aren't rendering time
spine table names the same way we do other table names. This fixes that
by using the `relation_name` field from the `PydanticNodeRelation`. This
field includes proper quoting that we need to use when rendering the
table name in SQL.