Bump DSI Version (#1876)
Version BUUUUUUUMP

This PR is just here to pull in the new transformations and such from dsi.  Early commits in this PR will be pinned to commits while I'm working on feeding https://github.com/dbt-labs/dbt-semantic-interfaces/pull/436 to the merge-asaurus.

Snapshot updates reflect the fact that now(*) we produce more metrics to replace measures, reflecting the changes in structure for the new user YAML.

(*) "now" is true, but ONLY IN TESTS.  Because of a transformation rule that exists only in MF right now, we don't use the list of transformations directly from dsi.  See upstack #1877 for the PR that brings these changes to prod.  OTH, tests DO run on the default list of metric transformations from dsi, so the snapshots needed more immediate updating.