Enable `DimensionPattern` without time dimensions (#1883)
There's a bug in MFS due to the `LooseTimeDimensionParameter`
workaround. This workaround was intended to allow you to pass a grain on
a categorical dimension without error in order to avoid breaking some
integrations. However, it also has the unintended consequence of not
raising an error if the grain you requested is not supported for the
selected time dimension. Adding this parameter will enable fixing this
issue in MFS. Draft PR
[here](https://github.com/dbt-labs/metricflow-server/pull/1618) to see
how this will be used.