Update `metricflow` to migrate `measures -> simple metrics` (#1889)
This PR contains the updates required for the `measure -> simple metric`
migration in `metricflow`. With the elimination of measures,
`MeasureSpec` was replaced by `SimpleMetricInputSpec` to represent the
state of a column before a simple metric is computed (e.g. with a time
spine join).

Significant changes were made to the dataflow plan builder to allow for
simple metrics to be used as building blocks for cumulative / conversion
metrics, which previously used measures as inputs.

There is currently a difference in filtering behavior (see
`test_join_to_time_spine_with_filter_not_in_group_by`) that needs
resolution (the filters to apply to the time spine before the join) and
currently produces a snapshot diff.