Sync DSI into MF (#1895)
We are consolidating `dbt-semantic-interfaces` into `metricflow`. This
is step one of that process. Note that this commit does not remove the
DSI dependency in MF or change the import paths to use the synced files,
that change will come later.
- Adds a script to copy DSI into MF, and update everything needed to
pass CI. This may need to be run multiple times before we are able to
deprecate the DSI repo.
- I ran the script, so this PR also includes the changes copied over
from DSI

I highly recommend reviewing by commit!