Fix optimizer bug for combining simple metrics (#1935)
## Context
There was a bug where 2 simple metrics were being joined together in a
derived metric and when both metrics has an alias, the optimizer ends up
not rendering the expr for one of the metric. This seemed to expose a
bug in the optimizer and this PR fixes that.

NOTE: Check first commit test case to see the bug