Render simple-metric inputs without a dunder prefix for the `WHERE` filter (#1941)
This PR updates the SQL rendering for the `WhereConstraintNode` to use a
subquery that generates column aliases without a dunder prefix for
simple metric inputs e.g. (`bookings` instead of `__bookings`).

This is added to handle a migration case - some customers have been
writing SQL strings in the filters that reference internal column names,
which is an unsupported use case. However, to reduce errors during
migration, this PR temporarily restores the previous behavior.

Due to the snapshot changes, please view by commit.