Selective aliasing for simple-metrics in `WHERE` filters (#1943)
This PR is an update to #1941 to handle an edge case with manifests that
contain a metric and an entity with the same name. Instead of aliasing
all simple-metrics, the change is to alias only a specific one that is
relevant to the containing branch in the dataflow plan. This limits
collisions in the `SELECT` statement that is generated from the branch.

Due to the large snapshot changes, please view by commit.