Replace generic mf_test_configuration with cli-specific fixture
The last remaining meaningful linkage between the CLI tests and
the broader tests_metricflow package was the mf_test_configuration
fixture, which provides a lot of SqlClient configuration parameters
and other similar things.

Since the CLI tests exclusively use tutorial configuration via a
dbt-duckdb integration most of these configuration constructs are
not used, so we make a simplified instance of the object in a
CLI-specific fixture and use that instead.