Move CLI tests to dbt-metricflow package
We are working to eliminate all dbt dependencies from the main
metricflow package in preparation for making metricflow itself
a required dependency for dbt-core. The only piece of the
metricflow package that requires a dbt dependency is the CLI
test suite, which we kept located with the MetricFlow tests
out of temporary convenience.

This moves all of the files for the CLI tests into their natural
home, a test directory inside of the dbt-metricflow package. While
a multi-package test configuration is a possibility it represents
a much larger overhaul, so we're just splitting the packages in a
more natural way instead.