Enable dbt-metricflow tests environment and link it to makefile
This makes the necessary changes for allowing developers to run
these tests, and fixes all of the broken imports and so forth
in the tests themselves.

Note we need to use the editable install otherwise import paths
break in the test runner if anybody moves a file in the metricflow
or metricflow-semantics packages. This slows down test runtime
considerably due to the added pip install step, so we designate
all CLI tests as "slow" in the makefile.