Fix logging setup in the CLI (#1932)
The `mf` CLI was not logging `INFO` messages to the log file. This was
traced to an imported module setting up the log configuration and
`logging.basicConfig` becoming a no-op if handlers were already added to
the root logger.

This PR updates the CLI's logging setup method to use
`logging.basicConfig(..., force=True)` to address the no-op case and
cleans up how the stream handler and the log file handler are set up.

A test is added separately
https://github.com/dbt-labs/metricflow/pull/1934 due to a required PR in
the middle.