Move CLI tests to dbt-metricflow package (#1948)
Move CLI tests to dbt-metricflow package

We are working to eliminate all dbt dependencies from the main
metricflow package in preparation for making metricflow itself
a required dependency for dbt-core. The only piece of the
metricflow package that requires a dbt dependency is the CLI
test suite, which we kept located with the MetricFlow tests
out of temporary convenience.

This moves all of the files for the CLI tests into their natural
home, a test directory inside of the dbt-metricflow package. While
a multi-package test configuration is a possibility it represents
a much larger overhaul, so we're just splitting the packages in a
more natural way instead.

This change also makes the necessary changes for allowing developers
to run these tests from their new location, and fixes all of the broken
imports and so forth in the tests themselves. Snapshots were ported via 
`git mv` rather than regeneration, so they should not have changed
 at all.

Note we need to use the editable install otherwise import paths
break in the test runner if anybody moves a file in the metricflow
or metricflow-semantics packages. This slows down test runtime
considerably due to the added pip install step, so we designate
all CLI tests as "slow" in the makefile.